rename _sv_parser to parser.py
