+++ src/sv_simpleparser/parser.py
@@ -20,6 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+"""SystemVerilog Parser."""
+
 import logging
 from dataclasses import dataclass
 from pathlib import Path
